+++ packages/blazor-workspace/NimbleBlazor/Source/TableColumn/EnumBase/NimbleTableColumnEnumBase.cs
@@ -1,5 +1,4 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using Microsoft.AspNetCore.Components;
+﻿using Microsoft.AspNetCore.Components;
 
 namespace NimbleBlazor;
 
@@ -49,8 +48,8 @@ public class NimbleTableColumnEnumBase<TKey> : NimbleTableColumn, IGroupableColu
     [Parameter]
     public bool? SortingDisabled { get; set; }
 
-    private readonly Type[] supportedCSharpNumberTypes = new[]
-    {
+    private readonly Type[] supportedCSharpNumberTypes =
+    [
         typeof(int),
         typeof(uint),
         typeof(short),
@@ -59,14 +58,14 @@ public class NimbleTableColumnEnumBase<TKey> : NimbleTableColumn, IGroupableColu
         typeof(sbyte),
         typeof(float),
         typeof(double)
-    };
+    ];
 
-    private readonly Type[] unsupportedCSharpNumberTypes = new[]
-    {
+    private readonly Type[] unsupportedCSharpNumberTypes =
+    [
         typeof(long),
         typeof(ulong),
         typeof(decimal)
-    };
+    ];
 
     protected string GetTKeyAsJSType()
     {