+++ packages/blazor-workspace/Tests/NimbleBlazor.Tests/Unit/Components/NimbleAnchorTabsTests.cs
@@ -1,4 +1,3 @@
-using System.Linq;
 using AngleSharp.Dom;
 using BlazorWorkspace.Testing.Unit;
 using Bunit;