+++ packages/blazor-workspace/Tests/NimbleBlazor.Tests/Unit/Components/NimbleTableTests.cs
@@ -1,8 +1,5 @@
-using System;
-using System.Collections.Generic;
 using System.Linq.Expressions;
 using System.Text.Json;
-using System.Threading.Tasks;
 using BlazorWorkspace.Testing.Unit;
 using Bunit;
 using Xunit;