+++ packages/blazor-workspace/Tests/NimbleBlazor.Tests/Unit/Components/NimbleTabsTests.cs
@@ -1,5 +1,3 @@
-using System.Linq;
-using System.Threading.Tasks;
 using AngleSharp.Dom;
 using BlazorWorkspace.Testing.Unit;
 using Bunit;