+++ packages/spright-components/src/chat/input/tests/chat-input.spec.ts
@@ -330,6 +330,18 @@ describe('ChatInput', () => {
             expect(spy).not.toHaveBeenCalled();
         });
 
+        it('via Enter with sendDisabled set triggers no send event', async () => {
+            const spy = jasmine.createSpy();
+            element.addEventListener('send', spy);
+            element.value = 'new value';
+            element.sendDisabled = true;
+            processUpdates();
+            await page.pressEnterKey();
+
+            expect(spy).not.toHaveBeenCalled();
+            expect(element.value).toBe('new value');
+        });
+
         it('Shift-Enter triggers no send event', async () => {
             const spy = jasmine.createSpy();
             element.addEventListener('send', spy);