+++ packages/blazor-workspace/SprightBlazor/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/spright-blazor
 
-<!-- This log was last generated on Mon, 15 Jun 2026 17:15:51 GMT and should not be manually modified. -->
+<!-- This log was last generated on Thu, 18 Jun 2026 00:03:55 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 4.16.2
+
+Thu, 18 Jun 2026 00:03:55 GMT
+
+### Patches
+
+- Updates from CA rule changes ([ni/nimble@894b5ac](https://github.com/ni/nimble/commit/894b5acee617ec104097c19bdc2f27daba66cdd4))
+
+## 4.16.1
+
+Wed, 17 Jun 2026 21:07:03 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v6.20.2
 
 ## 4.16.0